+++ src/gpjson/index/kernels/sharemem/combined-escape-carry-newline-count-index.cu
@@ -1,21 +1,23 @@
+#include <cstddef>
+
 namespace gpjson::index::kernels::sharemem {
 
 __global__ void combined_escape_carry_newline_count_index(const char *file,
-                                                          int fileSize,
+                                                          size_t fileSize,
                                                           char *escapeCarry,
                                                           int *newlineCount) {
-  int index = blockIdx.x * blockDim.x + threadIdx.x;
-  int stride = blockDim.x * gridDim.x;
+  size_t index = static_cast<size_t>(blockIdx.x) * blockDim.x + threadIdx.x;
+  size_t stride = static_cast<size_t>(blockDim.x) * gridDim.x;
 
-  int charsPerThread = (fileSize + stride - 1) / stride;
-  int bitmapAlignedCharsPerThread = ((charsPerThread + 64 - 1) / 64) * 64;
-  int start = index * bitmapAlignedCharsPerThread;
-  int end = start + bitmapAlignedCharsPerThread;
+  size_t charsPerThread = (fileSize + stride - 1) / stride;
+  size_t bitmapAlignedCharsPerThread = ((charsPerThread + 64 - 1) / 64) * 64;
+  size_t start = index * bitmapAlignedCharsPerThread;
+  size_t end = start + bitmapAlignedCharsPerThread;
 
   char carry = 0;
   int count = 0;
 
-  for (int i = start; i < end && i < fileSize; i += 1) {
+  for (size_t i = start; i < end && i < fileSize; i += 1) {
     char value = file[i];
     if (value == '\\') {
       carry = 1 ^ carry;
